Test if lock file is null
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@254 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/org/objectweb/proactive/core/jini/ServiceLocatorHelper.java
@@ -116,7 +116,9 @@ public class ServiceLocatorHelper implements DiscoveryListener {
     try {
     	getOrCreateServiceLocator();
     	//delete the lock file
-      if(jiniLockFile.exists()) jiniLockFile.delete();
+      if(jiniLockFile != null){
+        if(jiniLockFile.exists()) jiniLockFile.delete();
+      }
     } catch(java.io.IOException e) {
       e.printStackTrace();
       System.exit(1);
